church_data_pull now pulls all appropriate data and nests it within a dictionary
